refactor(parser): rename fields of the parser to improve usability
Creates the Enclosure enumerator to improve the specification used enclosures.

BREAKING CHANGE: Changes in the parser class:
- The field methodInstructionArgsOpeningEnclosure now becomes the
methodInstructionArgsEnclosure field, which receives a value of the
Enclosure enumerator.
- The field methodInstructionTargetsOpeningEnclosure now becomes the
methodInstructionTargetsEnclosure field, which receives a value of the
Enclosure enumerator.
- The static field DEFAULT_METHOD_INSTRUCTION_ARGS_OPENING_ENCLOSURE now
becomes the DEFAULT_METHOD_INSTRUCTION_ARGS_ENCLOSURE static field, which
has a value of the Enclosure enumerator.
- The static field DEFAULT_METHOD_INSTRUCTION_TARGETS_OPENING_ENCLOSURE
now becomes the DEFAULT_METHOD_INSTRUCTION_TARGETS_ENCLOSURE static field,
which has a value of the Enclosure enumerator.